+++ apps/E2E/src/ButtonV1/specs/ButtonV1.spec.win.ts
@@ -1,21 +1,19 @@
-import NavigateAppPage from '../../common/NavigateAppPage';
 import ButtonV1PageObject from '../pages/ButtonV1PageObject';
-import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT, BUTTON_A11Y_ROLE, Keys, Attribute } from '../../common/consts';
+import { BUTTON_A11Y_ROLE, Keys, Attribute } from '../../common/consts';
 import { BUTTON_ACCESSIBILITY_LABEL, BUTTON_TEST_COMPONENT_LABEL } from '../../ButtonLegacy/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('ButtonV1 Testing Initialization', () => {
   it('Wait for app load', async () => {
-    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
+    await ButtonV1PageObject.waitForInitialPageToDisplay();
+    expect(await ButtonV1PageObject.isInitialPageDisplayed()).toBeTruthy(ButtonV1PageObject.ERRORMESSAGE_APPLOAD);
   });
 
   it('Click and navigate to ButtonV1 test page', async () => {
     /* Click on component button to navigate to test page */
-    await NavigateAppPage.clickAndGoToButtonPage();
-    await ButtonV1PageObject.waitForPageDisplayed(PAGE_TIMEOUT);
+    await ButtonV1PageObject.navigateToPageAndLoadTests(true);
+    expect(await ButtonV1PageObject.isPageLoaded()).toBeTruthy(ButtonV1PageObject.ERRORMESSAGE_PAGELOAD);
 
-    await expect(await ButtonV1PageObject.isPageLoaded()).toBeTruthy(ButtonV1PageObject.ERRORMESSAGE_PAGELOAD);
     await expect(await ButtonV1PageObject.didAssertPopup()).toBeFalsy(ButtonV1PageObject.ERRORMESSAGE_ASSERT); // Ensure no asserts popped up
   });
 });