+++ apps/E2E/src/LinkV1/specs/LinkV1.spec.win.ts
@@ -1,21 +1,17 @@
-import NavigateAppPage from '../../common/NavigateAppPage';
 import LinkV1PageObject from '../pages/LinkV1PageObject';
 import { LINKV1_ACCESSIBILITY_LABEL } from '../consts';
-import { LINK_A11Y_ROLE, PAGE_TIMEOUT, BOOT_APP_TIMEOUT, Attribute, Keys } from '../../common/consts';
+import { LINK_A11Y_ROLE, Attribute, Keys } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('LinkV1 Testing Initialization', () => {
   it('Wait for app load', async () => {
-    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
+    await LinkV1PageObject.waitForInitialPageToDisplay();
+    expect(await LinkV1PageObject.isInitialPageDisplayed()).toBeTruthy(LinkV1PageObject.ERRORMESSAGE_APPLOAD);
   });
 
   it('Click and navigate to LinkV1 test page', async () => {
-    /* Click on component button to navigate to test page */
-    await NavigateAppPage.clickAndGoToLinkV1Page();
-    await LinkV1PageObject.waitForPageDisplayed(PAGE_TIMEOUT);
-
-    await expect(await LinkV1PageObject.isPageLoaded()).toBeTruthy(LinkV1PageObject.ERRORMESSAGE_PAGELOAD);
+    await LinkV1PageObject.navigateToPageAndLoadTests(true);
+    expect(await LinkV1PageObject.isPageLoaded()).toBeTruthy(LinkV1PageObject.ERRORMESSAGE_PAGELOAD);
 
     await expect(await LinkV1PageObject.didAssertPopup()).toBeFalsy(LinkV1PageObject.ERRORMESSAGE_ASSERT); // Ensure no asserts popped up
   });