+++ apps/fluent-tester/src/FluentTester.tsx
@@ -1,5 +1,5 @@
 import type { Theme } from '@fluentui-react-native/framework';
-import { Separator, Text } from '@fluentui/react-native';
+import { Separator, TextV1 as Text } from '@fluentui/react-native';
 import { ButtonV1 as Button } from '@fluentui-react-native/button';
 import { themedStyleSheet } from '@fluentui-react-native/themed-stylesheet';
 import * as React from 'react';