+++ apps/fluent-tester/src/TestComponents/Button/ButtonTest.tsx
@@ -61,15 +61,22 @@ const buttonSections: TestSection[] = [
         name: 'Deprecated Icon Button',
         component: ButtonIconTest_deprecated,
       },
-      {
-        name: 'Deprecated E2E Button Testing',
-        component: E2EButtonTest_deprecated,
-      },
     ],
   }),
   {
     name: 'E2E Button Testing',
     component: E2EButtonTest,
+  },
+];
+
+const e2eSections: TestSection[] = [
+  {
+    name: 'E2E Button Testing',
+    component: E2EButtonTest,
+  },
+  {
+    name: 'Deprecated E2E Button Testing',
+    component: E2EButtonTest_deprecated,
   },
 ];
 
@@ -87,5 +94,7 @@ export const ButtonTest: React.FunctionComponent = () => {
 
   const spec = 'https://github.com/microsoft/fluentui-react-native/blob/main/packages/components/Button/SPEC.md';
 
-  return <Test name="Button Test" description={description} spec={spec} sections={buttonSections} status={status} />;
+  return (
+    <Test name="Button Test" description={description} spec={spec} sections={buttonSections} status={status} e2eSections={e2eSections} />
+  );
 };