+++ apps/fluent-tester/src/TestComponents/LinkV1/LinkV1Test.tsx
@@ -169,7 +169,9 @@ const linkSections: TestSection[] = [
       component: CustomLinks,
     },
   }),
+];
 
+const e2eSections: TestSection[] = [
   {
     name: 'Link E2E Test',
     component: E2ELinkV1Test,
@@ -188,5 +190,5 @@ export const LinkV1Test: React.FunctionComponent = () => {
   const description =
     'With a Link, users can navigate to another page, window, or Help topic; display a definition; initiate a command; or choose an option. A Link indicates that it can be clicked, typically by being displayed using the visited or unvisited link system colors. Traditionally, Links are underlined as well, but that approach is often unnecessary and falling out of favor to reduce visual clutter.\n\nA Link is the lightest weight clickable control, and is often used to reduce the visual complexity of a design.';
 
-  return <Test name="LinkV1 Test" description={description} sections={linkSections} status={status}></Test>;
+  return <Test name="LinkV1 Test" description={description} sections={linkSections} status={status} e2eSections={e2eSections}></Test>;
 };