+++ apps/fluent-tester/src/theme/ThemePickers.ios.tsx
@@ -6,8 +6,6 @@ import { MenuView } from '@react-native-menu/menu';
 import { testerTheme } from './CustomThemes';
 import type { ThemeNames } from './applyTheme';
 import { themeChoices } from './applyTheme';
-import type { OfficeBrand } from './applyBrand';
-import { brandOptions } from './applyBrand';
 import type { Theme } from '@fluentui-react-native/framework';
 import { useTheme } from '@fluentui-react-native/framework';
 import { themedStyleSheet } from '@fluentui-react-native/themed-stylesheet';
@@ -38,10 +36,6 @@ export const themePickerStyles = themedStyleSheet((t: Theme) => {
 export const ThemePickers: React.FunctionComponent = () => {
   const themedStyles = themePickerStyles(useTheme());
 
-  const onBrandChange = React.useCallback((newBrand: string) => {
-    testerTheme.brand = newBrand as OfficeBrand;
-  }, []);
-
   const onThemeSelected = React.useCallback((newTheme: string) => {
     testerTheme.themeName = newTheme as ThemeNames;
   }, []);
@@ -50,12 +44,6 @@ export const ThemePickers: React.FunctionComponent = () => {
     id: themeChoice.value,
     title: themeChoice.label,
     state: testerTheme.themeName === themeChoice.value ? 'on' : 'off',
-  }));
-
-  const brandMenuOptions: MenuAction[] = brandOptions.map((brandOption) => ({
-    id: brandOption.value,
-    title: brandOption.label,
-    state: testerTheme.brand === brandOption.value ? 'on' : 'off',
   }));
 
   return (
@@ -68,15 +56,6 @@ export const ThemePickers: React.FunctionComponent = () => {
         actions={themeMenuOptions}
       >
         <Button appearance="subtle">Theme</Button>
-      </MenuView>
-      <MenuView
-        title="Brand"
-        onPressAction={({ nativeEvent }) => {
-          onBrandChange(nativeEvent.event);
-        }}
-        actions={brandMenuOptions}
-      >
-        <Button appearance="subtle">Brand</Button>
       </MenuView>
     </View>
   );