+++ apps/E2E/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/e2e-testing
 
-This log was last generated on Tue, 14 Feb 2023 02:41:04 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Feb 2023 18:14:14 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.20.1
+
+Fri, 17 Feb 2023 18:14:14 GMT
+
+### Patches
+
+- Add mobile e2e tests for Divider (winlarry@microsoft.com)
+
+## 1.20.0
+
+Thu, 16 Feb 2023 00:29:35 GMT
+
+### Minor changes
+
+- Add Divider E2E tests for desktop platforms (winlarry@microsoft.com)
+
+## 1.19.0
+
+Tue, 14 Feb 2023 21:33:24 GMT
+
+### Minor changes
+
+- Add retractable E2E sections to tester component pages (winlarry@microsoft.com)
 
 ## 1.18.0