+++ packages/experimental/Divider/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/divider
 
-This log was last generated on Sat, 11 Feb 2023 01:32:30 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Feb 2023 01:34:44 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.3
+
+Fri, 17 Feb 2023 01:34:44 GMT
+
+### Patches
+
+- Update SPEC.md documentation for Divider (winlarry@microsoft.com)
+
+## 0.2.2
+
+Wed, 15 Feb 2023 18:30:10 GMT
+
+### Patches
+
+- impl mobile divider (rohanpd.work@gmail.com)
+
+## 0.2.1
+
+Wed, 15 Feb 2023 00:51:10 GMT
+
+### Patches
+
+- Appearance prop correctly affects color (winlarry@microsoft.com)
 
 ## 0.2.0