+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Sat, 11 Feb 2023 01:32:30 GMT and should not be manually modified.
+This log was last generated on Tue, 14 Feb 2023 22:22:51 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.36.2
+
+Tue, 14 Feb 2023 22:22:51 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.32.32
+- Bump @fluentui-react-native/checkbox to v0.19.20
+- Bump @fluentui-react-native/menu-button to v0.10.31
 
 ## 0.36.1